repository and it's implementation

+++ app/src/main/java/com/project/coinswap/data/remote/dto/CurrencyDataDto.kt
@@ -32,6 +32,6 @@ data class CurrencyDataDto(
     val SGD: Double,
     val THB: Double,
     val TRY: Double,
-    val USD: Int,
+    val USD: Double,
     val ZAR: Double
 )
